backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
